LayoutTests/storage/indexeddb/transaction-state-active-after-creation-worker.html

<html>
<head>
<script src="../../resources/js-test.js"></script>
<script src="resources/shared.js"></script>
</head>
<body>
<script>
worker = startWorker('resources/transaction-state-active-after-creation.js');
</script>
</body>
</html>
